Initial map commit with all basic functionality
Implement:

	* map

	* markers

	* infowindow

	* search

All the basic functionality works. However, there was much agony over how googlemap should be placed in the Model View ViewModel structure. The map, the markers, and the infowindow are not data. Map is centered to this one location as long as the map is called Thong Lor map. So it's fixed for this app. Marker and InfoWindow are objects created by google.maps. They contain info and could set the DOM. Looks like as long as the code is clear and efficient, the positions won't matter.